requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
streamlit==1.41.1
langchain==0.3.18
langchain-google-genai==2.0.9
langchain-core==0.3.28
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.10.4
PyPDF2